requirements.txt
tensorflow==2.5.0
Flask
Pillow
scikit-learn
pandas
numpy
opencv-python
matplotlib